+++ sinocore/sinocore-fabric/src/main/java/games/moegirl/sinocraft/sinocore/utility/fabric/package-info.java
@@ -1,6 +1,6 @@
 @FieldsAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-package games.moegirl.sinocraft.sinocore.util.forge;
+package games.moegirl.sinocraft.sinocore.utility.fabric;
 
 import net.minecraft.FieldsAreNonnullByDefault;
 import net.minecraft.MethodsReturnNonnullByDefault;
