+++ sinocore/sinocore-forge/src/main/java/games/moegirl/sinocraft/sinocore/utility/forge/package-info.java
@@ -1,6 +1,6 @@
 @FieldsAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-package games.moegirl.sinocraft.sinocore.util.fabric;
+package games.moegirl.sinocraft.sinocore.utility.forge;
 
 import net.minecraft.FieldsAreNonnullByDefault;
 import net.minecraft.MethodsReturnNonnullByDefault;
