+++ sinocore/sinocore-neoforge/src/main/java/games/moegirl/sinocraft/sinocore/utility/neoforge/MenuHelperImpl.java
@@ -1,6 +1,6 @@
-package games.moegirl.sinocraft.sinocore.util.neoforge;
+package games.moegirl.sinocraft.sinocore.utility.neoforge;
 
-import games.moegirl.sinocraft.sinocore.util.MenuHelper;
+import games.moegirl.sinocraft.sinocore.utility.MenuHelper;
 import net.minecraft.server.level.ServerPlayer;
 import net.neoforged.neoforge.network.NetworkHooks;
