+++ sinocore/sinocore-neoforge/src/main/java/games/moegirl/sinocraft/sinocore/utility/neoforge/package-info.java
@@ -1,6 +1,6 @@
 @FieldsAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-package games.moegirl.sinocraft.sinocore.util.neoforge;
+package games.moegirl.sinocraft.sinocore.utility.neoforge;
 
 import net.minecraft.FieldsAreNonnullByDefault;
 import net.minecraft.MethodsReturnNonnullByDefault;
