+++ sinocore/src/main/java/games/moegirl/sinocraft/sinocore/mixin/AdvancementProviderMixin.java
@@ -1,6 +1,6 @@
 package games.moegirl.sinocraft.sinocore.mixin;
 
-import games.moegirl.sinocraft.sinocore.mixin_interfaces.injectable.IRenamedProvider;
+import games.moegirl.sinocraft.sinocore.mixin_interfaces.interfaces.IRenamedProvider;
 import net.minecraft.data.advancements.AdvancementProvider;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
