+++ sinocore/src/main/java/games/moegirl/sinocraft/sinocore/utility/EasyCast.java
@@ -1,4 +1,4 @@
-package games.moegirl.sinocraft.sinocore.util;
+package games.moegirl.sinocraft.sinocore.utility;
 
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
