+++ sinocore/src/main/java/games/moegirl/sinocraft/sinocore/utility/package-info.java
@@ -1,6 +1,6 @@
 @FieldsAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-package games.moegirl.sinocraft.sinocore.util;
+package games.moegirl.sinocraft.sinocore.utility;
 
 import net.minecraft.FieldsAreNonnullByDefault;
 import net.minecraft.MethodsReturnNonnullByDefault;
